feat: Only build the `nixosSystem` for the install phase
Gate `nixosSystem` build behind the `installPhase` flag.
This ensures that the `nixosSystem` is only built when really needed by
`nixos-anywhere`.

This is especially useful if `nixpkgs.hostPlatform` is not set yet and
building the derivation would error out.
The build can be split in to phases with `nixos-facter` being run in the
kexec phase.

Potential Drawbacks:

Since the `nixosSystem` is not built early in the invocation we can't
fail fast on errors anymore.
But this is only the case if the phases are already controlled from the
"outside" eg. if `nixos-anywhere` is called without the install phase
active.